Add new macros for convenience and move to seperate module
